Feat: move sorting button
